针对输入计算余弦函数

+++ src/main/java/func/CosineSimilarity.java
@@ -3,7 +3,7 @@ package func;
 import java.util.ArrayList;
 
 public class CosineSimilarity {
-    public double getSimilarity(ArrayList<Integer> words1, ArrayList<Integer> words2, int length1, int length2) {//传入两组向量与各自的长度
+    public double getSimilarity(ArrayList<Integer> words1, ArrayList<Integer> words2, double length1, double length2) {//传入两组向量与各自的长度
 
         double score = getSimilarityImpl(words1, words2, length1, length2);
 
@@ -13,7 +13,7 @@ public class CosineSimilarity {
         return score;
     }
 
-    public double getSimilarityImpl(ArrayList<Integer> words1, ArrayList<Integer> words2, int length1, int length2){
+    public double getSimilarityImpl(ArrayList<Integer> words1, ArrayList<Integer> words2, double length1, double length2){
         double num = 0;//点乘法结果
         for(int i = 0; i < words1.size(); i++)
         {
@@ -30,4 +30,5 @@ public class CosineSimilarity {
         }
         return (double)Math.sqrt((double)sum);
     }
+
 }
